internal/syscall/unix: fix GoString to avoid checkptr violations
Under the race detector, checkptr flags uses of unsafe.Slice that
result in slices that straddle multiple Go allocations.
Avoid that scenario by only calling unsafe.Slice on the range of data
known to be part of the valid string.

This fixes a failure on the darwin-.*-race builders introduced in
CL 449316.

Change-Id: Icd8851511e08199fe5aeeceb4358cc996eea94e8