cmd/link/internal/ppc64: fix trampoline reuse distance calculation
If a compatible trampoline has been inserted by a previously laid function
in the same section, and is known to be sufficiently close, it can be reused.

When testing if the trampoline can be reused, the addend of the direct call
should be ignored. It is already encoded in the trampoline. If the addend is
non-zero, and the target sufficiently far away, and just beyond direct call
reach, this may cause the trampoline to be incorrectly reused.

This was observed on go1.17.13 and openshift-installer commit f3c53b382
building in release mode with the following error:

github.com/aliyun/alibaba-cloud-sdk-go/services/cms.(*Client).DescribeMonitoringAgentAccessKeyWithChan.func1: direct call too far: runtime.duffzero+1f0-tramp0-1 -2000078

Fixes #56775

Change-Id: I54af957302506d4e3cd5d3121542c83fe980e912