go/types, types2: do not infer external unexported types
It should not be possible to infer a (named) type that cannot
be accessed explicitly if so desired, otherwise type inference
may enable code that is not otherwise possible (an explicit
non-goal of type inference).

This change checks that inferred named types are either exported
or local.

This change is not strictly backward-compatible but arguably
fixes a bug in the implementation.

For #56669.

Change-Id: Icb4b8741a6258b966a0a9b2483d6baef5f05d265