cmd/compile: add anchored version of SP
The SPanchored opcode is identical to SP, except that it takes a memory
argument so that it (and more importantly, anything that uses it)
must be scheduled at or after that memory argument.

This opcode ensures that a LEAQ of a variable gets scheduled after the
corresponding VARDEF for that variable.

It may lead to less CSE of LEAQ operations. TODO: figure out that effect.

WIP

Change-Id: Ib893562cd05369b91dd563b48fb83f5250950293