cmd/go/internal/modcmd: expand help text for go mod tidy
"go mod tidy" will ignore files with the build constraint "ignore". This is documented in https://go.dev/ref/mod#go-mod-tidy. While this URL in mentioned in the existing help text, the fact about the special handling of the `ignore` tag is significant enough to be mentioned explicitly in the help text. From the existing link it's not immediately obvious that this special behavior exists.

Take most of the relevant text from the online description and add it to the help text of "go mod tidy".

Fixes #54993

Change-Id: I806778a83d805440e2a19d2d71ddc2ed433032ac
GitHub-Last-Rev: dc73bc1f984fa0d1149b4ce7feca9cd3ca310c14
GitHub-Pull-Request: golang/go#54998