cmd/internal/obj/loong64, cmd/link/internal: switch to LoongArch ELF psABI v2 relocs
The LoongArch ELF psABI v2 [1] relocs are vastly simplified from the v1
which involved a stack machine for computing the reloc values, but the
details of PC-relative addressing are changed as well. Specifically, the
`pcaddu12i` instruction is substituted with the `pcalau12i`, which is
like arm64's `adrp` -- meaning the lower bits of a symbol's address now
have to be absolute and not PC-relative. However, the little bit of added
complexity apart, the obvious advantage is that only 1 reloc needs to be
emitted for every kind of external reloc we care about, which can mean
substantial space savings, and no open-coded stack ops has to remain any
more.

While at it, update the preset value for the output ELF's flags to
indicate the psABI update. This prevents legacy toolchains from seeing
the new reloc types too.

Updates #58784

[1]: https://loongson.github.io/LoongArch-Documentation/LoongArch-ELF-ABI-EN.html

Change-Id: I5c13bc710eaf58293a32e930dd33feff2ef14c28