encoding/json: implement more intelligent encoder buffer pooling
The current usage of sync.Pool is leaky because
it stores an arbitrary sized buffer into the pool.
However, sync.Pool assumes that all items in the pool
are interchangeable from a memory cost perspective.
Due to the unbounded size of a buffer that may be added,
it is possible for the pool to eventually pin
arbitrarily large amounts of memory in a live-lock situation.

This problem is hard to solve since we fundamentally
do not know up front how large the output will be
for a given Go object until we actually serialize it.
Thus, conventional approaches of tiered sync.Pools are not suitable
since that relies on knowing the buffer size up front.

A naive fix is to only pool buffers no larger than a fixed size,
but this hurts use cases that consistently marshals objects just above
the hard-coded fixed size.

Instead, this implementation tracks basic statistics about
buffer utilization on a per-buffer usage and
purges the buffer if it is under-utilized.
The implementation is simple, fast, and
easy to prove worst-case memory utilization performance.

The worst-case latency performance of this approach is when there is
a low-frequency stream of large objects concurrently serialized with
a high-frequency stream of small objects.
If the frequency of the large object stream is slow enough,
sufficiently large buffers will never be pooled.
However, this is arguably okay since the large objects evidently are not
common enough to justify keeping around idle memory resources.
Allocating fresh buffers for those large objects seems appropriate.

Performance numbers are noisy as results change by a few points across runs.
In general, the new approach adds a few nanoseconds to each call
(primarily the cost of executing integer arithmetic in putEncodeState).
Since each benchmark marshals the same object size,
the bulk of each benchmark is not changed since this pooling approach highly benefits
marshaling of similar-sized objects.

This implementation is derived from the experimental JSON implementation
in the github.com/go-json-experiment/json module and has seen some degree
of production usage at Tailscale Inc. without any notable negative effects.
That said, Tailscale's usage of JSON is hardly representative of how
JSON is used at other companies.

Fixes #27735
Updates #23199
Fixes kubernetes/kubernetes#114276

Change-Id: I4c96b0520f9e0be3f48025e5b6795fdcd3071761