cmd/go: make -gocoverdir flag an exposed flag for go test
This patch modifies the previously internal "-test.gocoverdir" flag to
be an exposed/supported flag for coverage testing. When a user passes
"-gocoverdir=<dir>" to a "go test" run, it instructs the Go command to
copy coverage profile data files generated by that run to the
specified directory. This allows users to collect the new-style
coverage data files generated from a "go test -cover" run, so as to
permit merging with other data files (for example, results from a
separate integration test run), something that users have requested.

Change-Id: Ied51d74e2455b321e806dd0325284de3bfc549ac