cmd/{go,cover,covdata}: fix 'package main' inconsistent handling
Fix a buglet in cmd/cover in how we handle package name/path for the
"go build -o foo.exe *.go" and "go run *.go" cases.

The go command assigns a dummy import path of "command-line-arguments"
to the main package built in these cases; rather than expose this
dummy to the user in coverage reports, the cover tool has a special
case (hack) that rewrites this package path to "main". The hack was
too general, however, and was rewriting the import path of all packages
with (p.name == "main") to an import path of "main".

This patch ensures that the hack applies only to the specific cases in
question; we don't want to rewrite if the main package has a real,
non-dummy import path.

Fixes #57169.

Change-Id: Ib6071db5e3485da3b8c26e16ef57f6fa1712402c